Update readme.md
Fix awesome badge rendering on Github.

There is some rendering issue with the `<img>` tag and the markdown for the awesome image. After trying a few methods to get the correct rendering, I have settled on this trivial change. The introduction of the single linefeed between the tag and the `# header`.