Add user-statistician GitHub Action
https://github.com/cicirello/user-statistician

A GitHub Action that generates SVG of detailed GitHub user activity for profile readmes. It includes international support with over 15 languages, and it is highly customizable. It runs entirely within GitHub Actions, such as on a cron schedule. The repository includes a directory with a few sample workflows to easily get users started.